fix precsision and conidtioning issues due to float32 and pinv/inv)

+++ utils/ICA.py
@@ -9,7 +9,7 @@ from statsmodels.stats.multitest import multipletests
 def ICA(whitened_data, output_dir="plots", random_state=None):
     ica = FastICA(whiten='unit-variance', random_state=random_state)
     # Takes in array-like of shape (n_samples, n_features) and returns ndarray of shape (n_samples, n_components)
-    spatial_components = ica.fit_transform(whitened_data.T).T
+    spatial_components = ica.fit_transform(whitened_data.T.astype(np.float64, copy=False)).T
 
     return spatial_components
